Analyzer clean and polish. Improve command line option handling, resolve crash for files that include effects undefined in the EP, tweak the build to use the expat library now present in the source tree. Review the python code.  Remove the gui parts because they weren't workign well cross platform and required numerous additions in addition to base python distribution. Retains the html generation code. Polished and cleaned this code to make it easiear to understand.  Slight modifications to requirement reporting in summary (report all of them for now). Add analyzerhtml_shortcut.html script to act as a simple wrapper around analyzerhtml.py that can be executed from a Windows desktop by simply dropping files on the icon. Added windist.py file to turn python utilities into standalone windows exe using py2exe.

+++ Utilities/AAFAnalyzer/EditProtocolTestImpl/DerivationChainStateMachine.cpp
@@ -309,7 +309,7 @@ bool DerivationChainStateMachine::IsKnownEvent( const aafUID_t event ) const
 AxString DerivationChainStateMachine::Indent()
 {
 
-  unsigned int l = _currentState.size() - 1;
+  size_t l = _currentState.size() - 1;
   AxString indent;
 
   for (; l > 0 ; l-- )
